insert cast to inptr_t to eliminate cast ptr to int warning     (That doesn't necessarily fix anything, tclcmdah.c is still a mess.      At least it doesn't break the game's "query" feature, which      was broken by different changes to tclcmdah.c in the recent lump set      of changes to tk and tcl to eliminate compile warnings.      That lump set was undone in order to find where the      problem was by committing them one by one. This was the one.)

+++ src/tcl/tclcmdah.c
@@ -886,7 +886,9 @@ Tcl_FormatCmd(dummy, interp, argc, argv)
 	    if (useTwoWords) {
 		sprintf(dst+dstSize, newFormat, twoWordValue);
 	    } else if (useShort) {
-	        int tmp = (int)oneWordValue;
+		/* sde warning was "cast from pointer to integer of different size" */
+		/* int tmp = (int)oneWordValue; */
+		int tmp = (int) (intptr_t) oneWordValue;
 		sprintf(dst+dstSize, newFormat, (short)tmp);
 	    } else {
 		sprintf(dst+dstSize, newFormat, oneWordValue);
